refactor(Nav): delegate Search-specific variables/hooks to SearchContext
