+++ client/src/Providers/index.ts
@@ -2,6 +2,7 @@ export { default as ToastProvider } from './ToastContext';
 export { default as AssistantsProvider } from './AssistantsContext';
 export * from './ChatContext';
 export * from './ToastContext';
+export * from './SearchContext';
 export * from './FileMapContext';
 export * from './AssistantsContext';
 export * from './AssistantsMapContext';
