+++ client/src/components/Nav/Nav.tsx
@@ -1,7 +1,6 @@
 import { useParams } from 'react-router-dom';
-import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { useRecoilValue } from 'recoil';
 import { useCallback, useEffect, useState, useMemo, memo } from 'react';
-import type { ConversationListResponse } from 'librechat-data-provider';
 import {
   useMediaQuery,
   useAuthContext,
@@ -10,9 +9,10 @@ import {
   useNavScrolling,
   useConversations,
 } from '~/hooks';
-import { useSearchInfiniteQuery, useConversationsInfiniteQuery } from '~/data-provider';
+import { useConversationsInfiniteQuery } from '~/data-provider';
 import { TooltipProvider, Tooltip } from '~/components/ui';
 import { Conversations } from '~/components/Conversations';
+import { useSearchContext } from '~/Providers';
 import { Spinner } from '~/components/svg';
 import SearchBar from './SearchBar';
 import NavToggle from './NavToggle';
@@ -47,24 +47,17 @@ const Nav = ({ navVisible, setNavVisible }) => {
     }
   }, [isSmallScreen]);
 
-  const [pageNumber, setPageNumber] = useState(1);
+  const { newConversation } = useConversation();
   const [showLoading, setShowLoading] = useState(false);
-
-  const searchQuery = useRecoilValue(store.searchQuery);
   const isSearchEnabled = useRecoilValue(store.isSearchEnabled);
-  const { newConversation, searchPlaceholderConversation } = useConversation();
 
   const { refreshConversations } = useConversations();
-  const setSearchResultMessages = useSetRecoilState(store.searchResultMessages);
+
+  const { pageNumber, searchQuery, setPageNumber, searchQueryRes } = useSearchContext();
 
   const { data, fetchNextPage, hasNextPage, isFetchingNextPage } = useConversationsInfiniteQuery(
     { pageNumber: pageNumber.toString(), isArchived: false },
     { enabled: isAuthenticated },
-  );
-
-  const searchQueryRes = useSearchInfiniteQuery(
-    { pageNumber: pageNumber.toString(), searchQuery: searchQuery, isArchived: false },
-    { enabled: isAuthenticated && !!searchQuery.length },
   );
 
   const { containerRef, moveToTop } = useNavScrolling({
@@ -80,21 +73,6 @@ const Nav = ({ navVisible, setNavVisible }) => {
       [],
     [data, searchQuery, searchQueryRes?.data],
   );
-
-  const onSearchSuccess = useCallback(({ data }: { data: ConversationListResponse }) => {
-    const res = data;
-    searchPlaceholderConversation();
-    setSearchResultMessages(res.messages);
-    /* disabled due recoil methods not recognized as state setters */
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []); // Empty dependency array
-
-  useEffect(() => {
-    //we use isInitialLoading here instead of isLoading because query is disabled by default
-    if (searchQueryRes.data) {
-      onSearchSuccess({ data: searchQueryRes.data.pages[0] });
-    }
-  }, [searchQueryRes.data, searchQueryRes.isInitialLoading, onSearchSuccess]);
 
   const clearSearch = () => {
     setPageNumber(1);
